Add download controller action

+++ app/Controller/Component/FaviconComponent.php
@@ -7,6 +7,9 @@ class FaviconComponent extends Component {
         parent::__construct($collection, $settings);
 
         $this->dir = TMP.'cakemarks-favicon';
+
+        $this->runs = Configure::read("favicon.runs");
+
     }
 
     public function get_favicon_filename($url) {
@@ -20,12 +23,10 @@ class FaviconComponent extends Component {
             mkdir($this->dir);
         }
 
-        if (is_file($target)) {
-            return $target;
-        }
-        else {
+        if (!is_file($target) && $this->runs > 0) {
             $favicon_url = 'http://g.etfv.co/'.$url;
             file_put_contents($target, file_get_contents($favicon_url));
+            $this->runs--;
             return $target;
         }
     }
